+++ src/cobib/config/config.py
@@ -352,13 +352,14 @@ class LabelSuffix(Enum):
 class DatabaseFormatConfig(_ConfigBase):
     """The `config.database.format` section."""
 
-    label_default: str = "{label}"
+    label_default: str = "{unidecode(label)}"
     """Specifies a default label format which will be used for database entry keys. The format of
     this option follows the f-string like formatting of modifications (see also the documentation
-    of the `cobib.commands.modify.ModifyCommand`). The default configuration value leaves the label
-    unchanged compared to the metadata provided by the source from which the entry gets added. A
-    more useful example is `"{author.split(' and ')[0].split()[-1]}{year}"` which takes the surname
-    of the first author and immediately appends the publication year."""
+    of the `cobib.commands.modify.ModifyCommand`). The default configuration value passes the
+    originally provided label through [text-unidecode](https://pypi.org/project/text-unidecode/)
+    which replaces all Unicode symbols with pure ASCII ones. A more useful example is
+    `"{unidecode(author.split(' and ')[0].split()[-1])}{year}"` which takes the surname of the first
+    author, replaces the Unicode characters and then immediately appends the publication year."""
     label_suffix: tuple[str, LabelSuffix] = field(default_factory=lambda: ("_", LabelSuffix.ALPHA))
     """Specifies the suffix format which is used to disambiguate labels if a conflict would occur.
     This option takes a tuple of length 2, where the first entry is the string separating the
@@ -372,10 +373,6 @@ class DatabaseFormatConfig(_ConfigBase):
     """Specifies whether latex warnings should not be ignored during the escaping of special
     characters. This is a simple option which gets passed on to the internally used `pylatexenc`
     library."""
-    unidecode_labels: bool = True
-    """Specifies whether the label of a database entry should undergo a "Unicode decoding". When
-    enabled (the default), this means the label will be passed through
-    [text-unidecode](https://pypi.org/project/text-unidecode/)."""
 
     @override
     def validate(self) -> None:
@@ -399,10 +396,6 @@ class DatabaseFormatConfig(_ConfigBase):
         self._assert(
             isinstance(self.suppress_latex_warnings, bool),
             "config.database.format.suppress_latex_warnings should be a boolean.",
-        )
-        self._assert(
-            isinstance(self.unidecode_labels, bool),
-            "config.database.format.unidecode_labels should be a boolean.",
         )